kmod: update entity status asynchronously
Since we can only change status of entity/module
in uprobe handler, let all other commands to update
their status if possible.

For example, When remove xx.patch, xx.patch -> removed.
But if the handler is not triggered, we still install
this patch. In the whole process, we only update the
status but never really to change it.

Signed-off-by: Longjun Luo <luolongjuna@gmail.com>