kmod: set patch sync rules for upatch
For upatch, all commands take effect for the
binary. But, the patch itself works for the
process.

To elimnate the gap, each process will sync
its status asynchronous.

Each time, we send commands for the binary.
The 'set_status' and 'set_patch' of the binary
will be updated.

And also, upatch will register the uprobe for
all changed functions. Within the handler,
each process will try to update its status
to the newest.

There are still many works need to do:
1) calculate the reference of each changed
function usage, so we can ummap it safely.
2) use jmp instruction to 'unregister' the
uprobe.

Signed-off-by: Longjun Luo <luolongjuna@gmail.com>