upatch: remove useless FDE from eh_frame section
Basically, eh_frame consists of CIE and FDE.
In upatch, we may remove useless symbol in the
relocation list of eh_frame. And it will make the
link to have the error - "no .eh_frame_hdr table..."

To solve this problem, we need to remove useless FDE
in the eh_frame at the same time.

Signed-off-by: Longjun Luo <luolongjun@huawei.com>