syscared: fix 'kmod patch conflicts with itself' issue
  We use a collision list to record symbol overriding. Each
patch symbol has it's own collision list, which is saved in
an IndexMap. Key of the IndexMap is 'old_addr' of a symbol.
But for kmod patch, the 'old_addr' would be always 0. Thus,
the first patched symbol in kmod would be treated as overrided
by following patch symbol.  To solve this, we changed the key
from 'old_addr' to 'name' of patch symbols.

Signed-off-by: renoseven <dev@renoseven.net>